Added information about eclipse setup to README.md.
When working with eclipse it might happen that the VM of a JRE is chosen instead of the VM of a JDK to run eclipse.
This might lead to issues when m2e tries to resolve the path to the tools.jar dependency.
If the VM is specified explicitly via the -vm parameter in the eclipse.ini things work smoothly.

Signed-off-by: Victor Toni <victor.toni@gmail.com>